requirements.txt
django
google-generativeai
python-dotenv
pyttsx3
SpeechRecognition
requests
sympy
google-api-python-client
google-auth-oauthlib
pillow
pytesseract
pywhatkit
wikipedia
geocoder
dateparser
langdetect
googletrans==4.0.0-rc1
pyjokes
pyaudio